#895: fix blueprint logging (#897)

+++ src/background/logging.ts
@@ -19,10 +19,10 @@ import { v4 as uuidv4 } from "uuid";
 import { liftBackground } from "@/background/protocol";
 import { rollbar } from "@/telemetry/rollbar";
 import { MessageContext, Logger as ILogger, SerializedError } from "@/core";
-import { JsonObject } from "type-fest";
+import { Except, JsonObject } from "type-fest";
 import { deserializeError, serializeError } from "serialize-error";
 import { DBSchema, openDB } from "idb/with-async-ittr";
-import { reverse, sortBy } from "lodash";
+import { reverse, sortBy, isEmpty } from "lodash";
 import { _getDNT } from "@/background/telemetry";
 import { isContentScript } from "webext-detect-page";
 import { readStorage, setStorage } from "@/chrome";
@@ -37,6 +37,7 @@ import { expectBackgroundPage } from "@/utils/expectContext";
 
 const STORAGE_KEY = "LOG";
 const ENTRY_OBJECT_STORE = "entries";
+const DB_VERSION_NUMBER = 2;
 
 export type MessageLevel = "trace" | "debug" | "info" | "warn" | "error";
 
@@ -65,38 +66,50 @@ interface LogDB extends DBSchema {
     indexes: {
       extensionPointId: string;
       extensionId: string;
+      blueprintId: string;
       blockId: string;
       serviceId: string;
       authId: string;
-      context: [string, string, string, string, string];
+      context: [string, string, string, string, string, string];
     };
   };
 }
 
-const indexKeys = [
+const indexKeys: Array<keyof Except<MessageContext, "deploymentId">> = [
   "extensionPointId",
   "extensionId",
+  "blueprintId",
   "blockId",
   "serviceId",
   "authId",
 ];
 
 async function getDB() {
-  return openDB<LogDB>(STORAGE_KEY, 1, {
+  return openDB<LogDB>(STORAGE_KEY, DB_VERSION_NUMBER, {
     upgrade(db) {
+      try {
+        // For now, just clear local logs whenever we need to upgrade the log database structure. There's no real use
+        // cases for looking at historic local logs
+        db.deleteObjectStore(ENTRY_OBJECT_STORE);
+        console.warn(
+          "Deleting object store %s for upgrade",
+          ENTRY_OBJECT_STORE
+        );
+      } catch {
+        // Not sure what will happen if the store doesn't exist (i.e., on initial install, so just NOP it
+      }
+
       // Create a store of objects
       const store = db.createObjectStore(ENTRY_OBJECT_STORE, {
         keyPath: "uuid",
       });
-      store.createIndex("extensionPointId", "context.extensionPointId", {
-        unique: false,
-      });
-      store.createIndex("extensionId", "context.extensionId", {
-        unique: false,
-      });
-      store.createIndex("blockId", "context.blockId", { unique: false });
-      store.createIndex("serviceId", "context.serviceId", { unique: false });
-      store.createIndex("authId", "context.authId", { unique: false });
+      // Create individual indexes
+      for (const indexKey of indexKeys) {
+        store.createIndex(indexKey, `context.${indexKey}`, {
+          unique: false,
+        });
+      }
+      // Create the joint index
       store.createIndex(
         "context",
         indexKeys.map((x) => `context.${x}`),
@@ -108,15 +121,19 @@ async function getDB() {
 
 export async function appendEntry(entry: LogEntry): Promise<void> {
   const db = await getDB();
-  await db.add("entries", entry);
+  await db.add(ENTRY_OBJECT_STORE, entry);
 }
 
-function makeMatchEntry(context: MessageContext): (entry: LogEntry) => boolean {
+function makeMatchEntry(
+  context: MessageContext = {}
+): (entry: LogEntry) => boolean {
   return (entry: LogEntry) =>
-    Object.entries(context ?? {}).every(
-      ([key, value]) =>
-        (entry.context ?? {})[key as keyof MessageContext] === value
-    );
+    indexKeys.every((key) => {
+      // eslint-disable-next-line security/detect-object-injection -- indexKeys is compile-time constant
+      const toMatch = context[key];
+      // eslint-disable-next-line security/detect-object-injection -- indexKeys is compile-time constant
+      return toMatch == null || entry.context?.[key] === toMatch;
+    });
 }
 
 export async function clearLogs(): Promise<void> {
@@ -130,8 +147,13 @@ export async function clearLog(context: MessageContext = {}): Promise<void> {
   const db = await getDB();
 
   const tx = db.transaction(ENTRY_OBJECT_STORE, "readwrite");
-  const match = makeMatchEntry(context);
 
+  if (isEmpty(context)) {
+    await tx.store.clear();
+    return;
+  }
+
+  const match = makeMatchEntry(context);
   for await (const cursor of tx.store) {
     if (match(cursor.value)) {
       await cursor.delete();
@@ -146,14 +168,15 @@ export async function getLog(
   const tx = db.transaction(ENTRY_OBJECT_STORE, "readonly");
   const match = makeMatchEntry(context);
 
-  const entries = [];
+  const matches = [];
   for await (const cursor of tx.store) {
     if (match(cursor.value)) {
-      entries.push(cursor.value);
+      matches.push(cursor.value);
     }
   }
 
-  return sortBy(reverse(entries), (x) => -Number.parseInt(x.timestamp, 10));
+  // Use both reverse and sortBy because we want insertion order if there's a tie in the timestamp
+  return sortBy(reverse(matches), (x) => -Number.parseInt(x.timestamp, 10));
 }
 
 function buildContext(
@@ -272,7 +295,7 @@ export class BackgroundLogger implements ILogger {
   }
 
   async error(error: unknown, data: JsonObject): Promise<void> {
-    console.error(`An error occurred: ${error}`, {
+    console.error(`An error occurred: %s`, getErrorMessage(error), {
       error,
       context: this.context,
       data,
